+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/drm/DefaultDrmSession.java
@@ -394,8 +394,13 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
       return true;
     } catch (NotProvisionedException e) {
       provisioningManager.provisionRequired(this);
-    } catch (Exception e) {
-      onError(e, DrmUtil.ERROR_SOURCE_EXO_MEDIA_DRM);
+    } catch (Exception | NoSuchMethodError e) {
+      // Work around b/291440132.
+      if (DrmUtil.isFailureToConstructNotProvisionedException(e)) {
+        provisioningManager.provisionRequired(this);
+      } else {
+        onError(e, DrmUtil.ERROR_SOURCE_EXO_MEDIA_DRM);
+      }
     }
 
     return false;
@@ -472,7 +477,7 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
     try {
       mediaDrm.restoreKeys(sessionId, offlineLicenseKeySetId);
       return true;
-    } catch (Exception e) {
+    } catch (Exception | NoSuchMethodError e) {
       onError(e, DrmUtil.ERROR_SOURCE_EXO_MEDIA_DRM);
     }
     return false;
@@ -492,7 +497,7 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
       currentKeyRequest = mediaDrm.getKeyRequest(scope, schemeDatas, type, keyRequestParameters);
       Util.castNonNull(requestHandler)
           .post(MSG_KEYS, Assertions.checkNotNull(currentKeyRequest), allowRetry);
-    } catch (Exception e) {
+    } catch (Exception | NoSuchMethodError e) {
       onKeysError(e, /* thrownByExoMediaDrm= */ true);
     }
   }
@@ -504,8 +509,8 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
     }
     currentKeyRequest = null;
 
-    if (response instanceof Exception) {
-      onKeysError((Exception) response, /* thrownByExoMediaDrm= */ false);
+    if (response instanceof Exception || response instanceof NoSuchMethodError) {
+      onKeysError((Throwable) response, /* thrownByExoMediaDrm= */ false);
       return;
     }
 
@@ -526,7 +531,7 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
         state = STATE_OPENED_WITH_KEYS;
         dispatchEvent(DrmSessionEventListener.EventDispatcher::drmKeysLoaded);
       }
-    } catch (Exception e) {
+    } catch (Exception | NoSuchMethodError e) {
       onKeysError(e, /* thrownByExoMediaDrm= */ true);
     }
   }
@@ -538,8 +543,12 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
     }
   }
 
-  private void onKeysError(Exception e, boolean thrownByExoMediaDrm) {
-    if (e instanceof NotProvisionedException) {
+  /**
+   * @param e Must be an instance of either {@link Exception} or {@link Error}.
+   */
+  private void onKeysError(Throwable e, boolean thrownByExoMediaDrm) {
+    if (e instanceof NotProvisionedException
+        || DrmUtil.isFailureToConstructNotProvisionedException(e)) {
       provisioningManager.provisionRequired(this);
     } else {
       onError(
@@ -550,11 +559,24 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
     }
   }
 
-  private void onError(Exception e, @DrmUtil.ErrorSource int errorSource) {
+  /**
+   * @param e Must be an instance of either {@link Exception} or {@link Error}.
+   */
+  private void onError(Throwable e, @DrmUtil.ErrorSource int errorSource) {
     lastException =
         new DrmSessionException(e, DrmUtil.getErrorCodeForMediaDrmException(e, errorSource));
     Log.e(TAG, "DRM session error", e);
-    dispatchEvent(eventDispatcher -> eventDispatcher.drmSessionManagerError(e));
+    if (e instanceof Exception) {
+      dispatchEvent(eventDispatcher -> eventDispatcher.drmSessionManagerError((Exception) e));
+    } else if (e instanceof Error) {
+      // Re-throw all Error types except a NoSuchMethodError caused by b/291440132.
+      if (!DrmUtil.isFailureToConstructResourceBusyException(e)
+          && !DrmUtil.isFailureToConstructNotProvisionedException(e)) {
+        throw (Error) e;
+      }
+    } else {
+      throw new IllegalStateException("Unexpected Throwable subclass", e);
+    }
     if (state != STATE_OPENED_WITH_KEYS) {
       state = STATE_ERROR;
     }