+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/drm/DefaultDrmSessionManagerTest.java
@@ -40,6 +40,7 @@ import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.robolectric.annotation.Config;
 import org.robolectric.shadows.ShadowLooper;
 
 /** Tests for {@link DefaultDrmSessionManager} and {@link DefaultDrmSession}. */
@@ -258,6 +259,21 @@ public class DefaultDrmSessionManagerTest {
 
   @Test(timeout = 10_000)
   public void maxConcurrentSessionsExceeded_allKeepAliveSessionsEagerlyReleased() throws Exception {
+    maxConcurrentSessionsExceededAllKeepAliveSessionsEagerlyReleased(
+        /* throwNoSuchMethodErrorForResourceBusy= */ false);
+  }
+
+  /** Testing workarounds for b/291440132. */
+  @Config(sdk = 34)
+  @Test(timeout = 10_000)
+  public void maxConcurrentSessionsExceeded_allKeepAliveSessionsEagerlyReleased_noSuchMethodError()
+      throws Exception {
+    maxConcurrentSessionsExceededAllKeepAliveSessionsEagerlyReleased(
+        /* throwNoSuchMethodErrorForResourceBusy= */ true);
+  }
+
+  private static void maxConcurrentSessionsExceededAllKeepAliveSessionsEagerlyReleased(
+      boolean throwNoSuchMethodErrorForResourceBusy) {
     ImmutableList<DrmInitData.SchemeData> secondSchemeDatas =
         ImmutableList.of(DRM_SCHEME_DATAS.get(0).copyWithData(TestUtil.createByteArray(4, 5, 6)));
     FakeExoMediaDrm.LicenseServer licenseServer =
@@ -267,7 +283,13 @@ public class DefaultDrmSessionManagerTest {
     DrmSessionManager drmSessionManager =
         new DefaultDrmSessionManager.Builder()
             .setUuidAndExoMediaDrmProvider(
-                DRM_SCHEME_UUID, uuid -> new FakeExoMediaDrm(/* maxConcurrentSessions= */ 1))
+                DRM_SCHEME_UUID,
+                uuid ->
+                    new FakeExoMediaDrm.Builder()
+                        .setMaxConcurrentSessions(1)
+                        .throwNoSuchMethodErrorForProvisioningAndResourceBusy(
+                            throwNoSuchMethodErrorForResourceBusy)
+                        .build())
             .setSessionKeepaliveMs(10_000)
             .setMultiSession(true)
             .build(/* mediaDrmCallback= */ licenseServer);
@@ -298,6 +320,23 @@ public class DefaultDrmSessionManagerTest {
   @Test(timeout = 10_000)
   public void maxConcurrentSessionsExceeded_allPreacquiredAndKeepaliveSessionsEagerlyReleased()
       throws Exception {
+    maxConcurrentSessionsExceededAllPreacquiredAndKeepaliveSessionsEagerlyReleased(
+        /* throwNoSuchMethodErrorForResourceBusy= */ false);
+  }
+
+  /** Testing workarounds for b/291440132. */
+  @Config(sdk = 34)
+  @Test(timeout = 10_000)
+  public void
+      maxConcurrentSessionsExceeded_allPreacquiredAndKeepaliveSessionsEagerlyReleased_noSuchMethodError()
+          throws Exception {
+    maxConcurrentSessionsExceededAllPreacquiredAndKeepaliveSessionsEagerlyReleased(
+        /* throwNoSuchMethodErrorForResourceBusy= */ true);
+  }
+
+  private static void
+      maxConcurrentSessionsExceededAllPreacquiredAndKeepaliveSessionsEagerlyReleased(
+          boolean throwNoSuchMethodErrorForResourceBusy) {
     ImmutableList<DrmInitData.SchemeData> secondSchemeDatas =
         ImmutableList.of(DRM_SCHEME_DATAS.get(0).copyWithData(TestUtil.createByteArray(4, 5, 6)));
     FakeExoMediaDrm.LicenseServer licenseServer =
@@ -308,7 +347,12 @@ public class DefaultDrmSessionManagerTest {
         new DefaultDrmSessionManager.Builder()
             .setUuidAndExoMediaDrmProvider(
                 DRM_SCHEME_UUID,
-                uuid -> new FakeExoMediaDrm.Builder().setMaxConcurrentSessions(1).build())
+                uuid ->
+                    new FakeExoMediaDrm.Builder()
+                        .setMaxConcurrentSessions(1)
+                        .throwNoSuchMethodErrorForProvisioningAndResourceBusy(
+                            throwNoSuchMethodErrorForResourceBusy)
+                        .build())
             .setSessionKeepaliveMs(10_000)
             .setMultiSession(true)
             .build(/* mediaDrmCallback= */ licenseServer);
@@ -606,6 +650,22 @@ public class DefaultDrmSessionManagerTest {
   @Test
   public void
       deviceNotProvisioned_exceptionThrownFromOpenSession_provisioningDoneAndOpenSessionRetried() {
+    deviceNotProvisionedExceptionThrownFromOpenSessionProvisioningDoneAndOpenSessionRetried(
+        /* throwNoSuchMethodErrorForNotProvisioned= */ false);
+  }
+
+  /** Testing workarounds for b/291440132. */
+  @Config(sdk = 34)
+  @Test
+  public void
+      deviceNotProvisioned_exceptionThrownFromOpenSession_provisioningDoneAndOpenSessionRetried_noSuchMethodError() {
+    deviceNotProvisionedExceptionThrownFromOpenSessionProvisioningDoneAndOpenSessionRetried(
+        /* throwNoSuchMethodErrorForNotProvisioned= */ true);
+  }
+
+  private static void
+      deviceNotProvisionedExceptionThrownFromOpenSessionProvisioningDoneAndOpenSessionRetried(
+          boolean throwNoSuchMethodErrorForNotProvisioned) {
     FakeExoMediaDrm.LicenseServer licenseServer =
         FakeExoMediaDrm.LicenseServer.allowingSchemeDatas(DRM_SCHEME_DATAS);
 
@@ -613,7 +673,12 @@ public class DefaultDrmSessionManagerTest {
         new DefaultDrmSessionManager.Builder()
             .setUuidAndExoMediaDrmProvider(
                 DRM_SCHEME_UUID,
-                uuid -> new FakeExoMediaDrm.Builder().setProvisionsRequired(1).build())
+                uuid ->
+                    new FakeExoMediaDrm.Builder()
+                        .setProvisionsRequired(1)
+                        .throwNoSuchMethodErrorForProvisioningAndResourceBusy(
+                            throwNoSuchMethodErrorForNotProvisioned)
+                        .build())
             .build(/* mediaDrmCallback= */ licenseServer);
     drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     drmSessionManager.prepare();
@@ -635,6 +700,22 @@ public class DefaultDrmSessionManagerTest {
   @Test
   public void
       deviceNotProvisioned_exceptionThrownFromGetKeyRequest_provisioningDoneAndOpenSessionRetried() {
+    deviceNotProvisionedExceptionThrownFromGetKeyRequestProvisioningDoneAndOpenSessionRetried(
+        /* throwNoSuchMethodErrorForNotProvisioned= */ false);
+  }
+
+  /** Testing workarounds for b/291440132. */
+  @Config(sdk = 34)
+  @Test
+  public void
+      deviceNotProvisioned_exceptionThrownFromGetKeyRequest_provisioningDoneAndOpenSessionRetried_noSuchMethodError() {
+    deviceNotProvisionedExceptionThrownFromGetKeyRequestProvisioningDoneAndOpenSessionRetried(
+        /* throwNoSuchMethodErrorForNotProvisioned= */ true);
+  }
+
+  private static void
+      deviceNotProvisionedExceptionThrownFromGetKeyRequestProvisioningDoneAndOpenSessionRetried(
+          boolean throwNoSuchMethodErrorForNotProvisioned) {
     FakeExoMediaDrm.LicenseServer licenseServer =
         FakeExoMediaDrm.LicenseServer.allowingSchemeDatas(DRM_SCHEME_DATAS);
 
@@ -646,6 +727,8 @@ public class DefaultDrmSessionManagerTest {
                     new FakeExoMediaDrm.Builder()
                         .setProvisionsRequired(1)
                         .throwNotProvisionedExceptionFromGetKeyRequest()
+                        .throwNoSuchMethodErrorForProvisioningAndResourceBusy(
+                            throwNoSuchMethodErrorForNotProvisioned)
                         .build())
             .build(/* mediaDrmCallback= */ licenseServer);
     drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
@@ -665,6 +748,21 @@ public class DefaultDrmSessionManagerTest {
 
   @Test
   public void deviceNotProvisioned_doubleProvisioningHandledAndOpenSessionRetried() {
+    deviceNotProvisionedDoubleProvisioningHandledAndOpenSessionRetried(
+        /* throwNoSuchMethodErrorForNotProvisioned= */ false);
+  }
+
+  /** Testing workarounds for b/291440132. */
+  @Config(sdk = 34)
+  @Test
+  public void
+      deviceNotProvisioned_doubleProvisioningHandledAndOpenSessionRetried_noSuchMethodError() {
+    deviceNotProvisionedDoubleProvisioningHandledAndOpenSessionRetried(
+        /* throwNoSuchMethodErrorForNotProvisioned= */ true);
+  }
+
+  private static void deviceNotProvisionedDoubleProvisioningHandledAndOpenSessionRetried(
+      boolean throwNoSuchMethodErrorForNotProvisioned) {
     FakeExoMediaDrm.LicenseServer licenseServer =
         FakeExoMediaDrm.LicenseServer.allowingSchemeDatas(DRM_SCHEME_DATAS);
 
@@ -672,7 +770,12 @@ public class DefaultDrmSessionManagerTest {
         new DefaultDrmSessionManager.Builder()
             .setUuidAndExoMediaDrmProvider(
                 DRM_SCHEME_UUID,
-                uuid -> new FakeExoMediaDrm.Builder().setProvisionsRequired(2).build())
+                uuid ->
+                    new FakeExoMediaDrm.Builder()
+                        .setProvisionsRequired(2)
+                        .throwNoSuchMethodErrorForProvisioningAndResourceBusy(
+                            throwNoSuchMethodErrorForNotProvisioned)
+                        .build())
             .build(/* mediaDrmCallback= */ licenseServer);
     drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     drmSessionManager.prepare();
@@ -693,6 +796,20 @@ public class DefaultDrmSessionManagerTest {
 
   @Test
   public void keyResponseIndicatesProvisioningRequired_provisioningDone() {
+    keyResponseIndicatesProvisioningRequiredProvisioningDone(
+        /* throwNoSuchMethodErrorForNotProvisioned= */ false);
+  }
+
+  /** Testing workarounds for b/291440132. */
+  @Config(sdk = 34)
+  @Test
+  public void keyResponseIndicatesProvisioningRequired_provisioningDone_noSuchMethodError() {
+    keyResponseIndicatesProvisioningRequiredProvisioningDone(
+        /* throwNoSuchMethodErrorForNotProvisioned= */ true);
+  }
+
+  private static void keyResponseIndicatesProvisioningRequiredProvisioningDone(
+      boolean throwNoSuchMethodErrorForNotProvisioned) {
     FakeExoMediaDrm.LicenseServer licenseServer =
         FakeExoMediaDrm.LicenseServer.requiringProvisioningThenAllowingSchemeDatas(
             DRM_SCHEME_DATAS);
@@ -700,7 +817,12 @@ public class DefaultDrmSessionManagerTest {
     DefaultDrmSessionManager drmSessionManager =
         new DefaultDrmSessionManager.Builder()
             .setUuidAndExoMediaDrmProvider(
-                DRM_SCHEME_UUID, uuid -> new FakeExoMediaDrm.Builder().build())
+                DRM_SCHEME_UUID,
+                uuid ->
+                    new FakeExoMediaDrm.Builder()
+                        .throwNoSuchMethodErrorForProvisioningAndResourceBusy(
+                            throwNoSuchMethodErrorForNotProvisioned)
+                        .build())
             .build(/* mediaDrmCallback= */ licenseServer);
     drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     drmSessionManager.prepare();
@@ -719,10 +841,29 @@ public class DefaultDrmSessionManagerTest {
 
   @Test
   public void provisioningUndoneWhileManagerIsActive_deviceReprovisioned() {
+    provisioningUndoneWhileManagerIsActiveDeviceReprovisioned(
+        /* throwNoSuchMethodErrorForNotProvisioned= */ false);
+  }
+
+  /** Testing workarounds for b/291440132. */
+  @Config(sdk = 34)
+  @Test
+  public void provisioningUndoneWhileManagerIsActive_deviceReprovisioned_noSuchMethodError() {
+    provisioningUndoneWhileManagerIsActiveDeviceReprovisioned(
+        /* throwNoSuchMethodErrorForNotProvisioned= */ true);
+  }
+
+  private static void provisioningUndoneWhileManagerIsActiveDeviceReprovisioned(
+      boolean throwNoSuchMethodErrorForNotProvisioned) {
     FakeExoMediaDrm.LicenseServer licenseServer =
         FakeExoMediaDrm.LicenseServer.allowingSchemeDatas(DRM_SCHEME_DATAS);
 
-    FakeExoMediaDrm mediaDrm = new FakeExoMediaDrm.Builder().setProvisionsRequired(2).build();
+    FakeExoMediaDrm mediaDrm =
+        new FakeExoMediaDrm.Builder()
+            .setProvisionsRequired(2)
+            .throwNoSuchMethodErrorForProvisioningAndResourceBusy(
+                throwNoSuchMethodErrorForNotProvisioned)
+            .build();
     DefaultDrmSessionManager drmSessionManager =
         new DefaultDrmSessionManager.Builder()
             .setUuidAndExoMediaDrmProvider(DRM_SCHEME_UUID, new AppManagedProvider(mediaDrm))